+++ apps/ui-dev/component/PageTools.jsx
@@ -1,43 +1,29 @@
-import { classnames, useStoreSet } from "@elf-framework/sapa";
-import { Button, Tools } from "@elf-framework/ui";
+import MenuFilled from "@elf-framework/icon/MenuFilled";
+import SearchFilled from "@elf-framework/icon/SearchFilled";
+import { classnames, useSetStoreValue } from "@elf-framework/sapa";
+import { Button } from "@elf-framework/ui";
 
 import mainMenus from "../constants/main-menus";
 import "./PageTools.scss";
 import { PeerView } from "./utils/peer-view/PeerView";
 import { ThemeButton } from "./utils/theme-button/ThemeButton";
 
-export function PageTools({ menu }) {
+export function PageTools() {
   const pathname = location.pathname;
+
+  const setShowMobileMenu = useSetStoreValue("show.mobile.menu");
+  const setOpenSearchView = useSetStoreValue("open.search.view");
 
   return (
     <div class="page-tools">
       <div class="sm">
-        <Tools
-          items={[
-            {
-              type: "menu",
-              title: "Menu",
-              items: [
-                {
-                  type: "item",
-                  title: "Documents",
-                  items: [{ type: "group", title: "Docs" }, ...mainMenus],
-                },
-                "-",
-                ...menu.map((it) => {
-                  if (typeof it === "string") {
-                    return { type: "group", title: it };
-                  }
-                  return {
-                    type: "link",
-                    selected: pathname.startsWith(it.category),
-                    ...it,
-                  };
-                }),
-              ],
-            },
-          ]}
-        />
+        <Button iconOnly quiet onClick={() => setOpenSearchView(true)}>
+          <SearchFilled />
+        </Button>
+
+        <Button iconOnly quiet onClick={() => setShowMobileMenu(true)}>
+          <MenuFilled />
+        </Button>
       </div>
       <div class="lg">
         <div>
@@ -67,7 +53,7 @@ export function PageTools({ menu }) {
         >
           <Button
             onClick={() => {
-              useStoreSet("open.search.view", true);
+              setOpenSearchView(true);
             }}
           >
             Open search