+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @elf-framework/ui
 
+## 0.0.146
+
+### Patch Changes
+
+- modify index.d.ts
+
 ## 0.0.140
 
 ### Patch Changes